refactor: colocate virtual chunk access logic in chunk_access module
