+++ pyproject.toml
@@ -131,7 +131,7 @@ ignore = [
 
 [tool.ruff.lint.per-file-ignores]
 "tests/test_app.py" = ["D1"]
-"tests/test_reader_credentials.py" = ["D1"]
+"tests/test_chunk_access.py" = ["D1"]
 "tests/fixtures/generate_test_zarr.py" = ["D1"]
 "tests/fixtures/generate_test_reference.py" = ["D1"]
