+++ src/titiler/multidim/chunk_access.py
@@ -1,0 +1,189 @@
+"""Typed configuration and icechunk credential construction for virtual chunk access.
+
+icechunk is imported inside the functions that need it, never at module
+level: settings.py imports this module's models, and settings.py must stay
+importable in the CDK deployment environment (`uv sync --only-group
+deployment`), which does not install icechunk.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any, Dict, Mapping, Optional, Union
+from urllib.parse import urlparse
+
+from pydantic import BaseModel, ConfigDict, model_validator
+
+if TYPE_CHECKING:
+    import icechunk
+
+
+class _CloudChunkAccess(BaseModel):
+    """Options shared by the cloud (s3/gcs/azure) virtual chunk entries.
+
+    Each field mirrors the JSON-expressible keyword of the matching icechunk
+    credential builder; only fields the user set are forwarded to it.
+    Credential use is opt-in: an entry must select an access mode explicitly
+    (anonymous where the builder supports it, from_env, or explicit credential
+    fields), so an empty entry cannot silently grant the service's own ambient
+    credentials.
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    from_env: bool | None = None
+
+    @model_validator(mode="after")
+    def _require_explicit_access_mode(self) -> "_CloudChunkAccess":
+        # icechunk's builders fall back to FromEnv (the service's own ambient
+        # credentials) when called with no arguments; requiring an explicit
+        # mode keeps that a deliberate grant instead of a default
+        static_fields = {
+            name
+            for name in self.model_fields_set
+            if name not in ("anonymous", "from_env") and getattr(self, name) is not None
+        }
+        if getattr(self, "anonymous", None) or self.from_env or static_fields:
+            return self
+        raise ValueError(
+            "credential use is opt-in: set 'from_env': true, 'anonymous': true "
+            "(s3/gcs only), or explicit credential fields"
+        )
+
+
+class S3ChunkAccess(_CloudChunkAccess):
+    """Options for an s3:// virtual chunk entry (icechunk.s3_credentials)."""
+
+    anonymous: bool | None = None
+    access_key_id: str | None = None
+    secret_access_key: str | None = None
+    session_token: str | None = None
+
+    def to_credential(self) -> icechunk.AnyS3Credential:
+        """Build the icechunk credential from the explicitly-set fields."""
+        import icechunk
+
+        return icechunk.s3_credentials(**self.model_dump(exclude_unset=True))
+
+
+class GcsChunkAccess(_CloudChunkAccess):
+    """Options for a gs://gcs:// virtual chunk entry (icechunk.gcs_credentials)."""
+
+    anonymous: bool | None = None
+    service_account_file: str | None = None
+    service_account_key: str | None = None
+    application_credentials: str | None = None
+    bearer_token: str | None = None
+
+    def to_credential(self) -> icechunk.AnyGcsCredential:
+        """Build the icechunk credential from the explicitly-set fields."""
+        import icechunk
+
+        return icechunk.gcs_credentials(**self.model_dump(exclude_unset=True))
+
+
+class AzureChunkAccess(_CloudChunkAccess):
+    """Options for an az://azure:// virtual chunk entry (icechunk.azure_credentials).
+
+    icechunk has no anonymous Azure credential variant, so unlike s3/gcs there
+    is no anonymous field here.
+    """
+
+    access_key: str | None = None
+    sas_token: str | None = None
+    bearer_token: str | None = None
+
+    def to_credential(self) -> icechunk.AnyAzureCredential:
+        """Build the icechunk credential from the explicitly-set fields."""
+        import icechunk
+
+        return icechunk.azure_credentials(**self.model_dump(exclude_unset=True))
+
+
+AnyChunkAccess = Union[S3ChunkAccess, GcsChunkAccess, AzureChunkAccess]
+
+# raw option dicts (from a caller) or parsed models (from settings)
+ChunkAccessMapping = Mapping[str, Union[Mapping[str, Any], AnyChunkAccess]]
+
+_SCHEME_MODELS: Dict[str, type[AnyChunkAccess]] = {
+    "s3": S3ChunkAccess,
+    "gs": GcsChunkAccess,
+    "gcs": GcsChunkAccess,
+    "az": AzureChunkAccess,
+    "azure": AzureChunkAccess,
+}
+
+
+def parse_chunk_access(
+    mapping: Mapping[str, Union[Mapping[str, Any], AnyChunkAccess]] | None,
+) -> Dict[str, AnyChunkAccess]:
+    """Parse a virtual chunk access config into per-scheme option models.
+
+    Each entry is dispatched by the URL scheme of its container prefix.
+    Already-parsed models pass through unchanged when they match their
+    prefix's scheme; a mismatched model, "file://" (must never read the
+    server's local filesystem), and unknown schemes are rejected.
+    """
+    parsed: Dict[str, AnyChunkAccess] = {}
+    for prefix, options in (mapping or {}).items():
+        scheme = urlparse(prefix).scheme
+        model = _SCHEME_MODELS.get(scheme)
+        if model is None:
+            if scheme == "file":
+                raise ValueError(
+                    f"refusing to authorize local filesystem virtual chunk access for {prefix!r}; "
+                    "virtual chunks must not read the server's local filesystem"
+                )
+            raise ValueError(
+                f"unsupported scheme {scheme!r} for virtual chunk entry {prefix!r}"
+            )
+        # urlparse lowercases the scheme, but icechunk matches container
+        # prefixes character for character, so an entry spelled 'S3://…'
+        # would validate here yet never match at request time
+        if not prefix.startswith(f"{scheme}://"):
+            raise ValueError(
+                f"virtual chunk entry {prefix!r} must begin with lowercase "
+                f"'{scheme}://'; icechunk matches container prefixes by exact "
+                "string, so other spellings are silently ignored"
+            )
+        if isinstance(options, BaseModel):
+            if not isinstance(options, model):
+                raise ValueError(
+                    f"virtual chunk entry {prefix!r} expects {model.__name__}, "
+                    f"got {type(options).__name__}"
+                )
+            parsed[prefix] = options
+            continue
+        try:
+            parsed[prefix] = model.model_validate(options)
+        except ValueError as e:
+            raise ValueError(
+                f"invalid options for virtual chunk entry {prefix!r}: {e}"
+            ) from e
+    return parsed
+
+
+def build_virtual_chunk_access(
+    authorize_virtual_chunk_access: Optional[ChunkAccessMapping],
+) -> Optional[Dict[str, Optional[icechunk.AnyCredential]]]:
+    """Translate virtual chunk access settings into icechunk credentials.
+
+    Entries are parsed into per-scheme option models by parse_chunk_access
+    (which also rejects file://, unknown schemes, and unrecognized options),
+    then each entry builds its own icechunk credential via to_credential().
+
+    Args:
+        authorize_virtual_chunk_access: Mapping of container URL prefix to
+            access options (raw dicts or parsed models).
+
+    Returns:
+        The mapping for Repository.open(authorize_virtual_chunk_access), or
+        None when no entries are configured.
+    """
+    import icechunk
+
+    entries = parse_chunk_access(authorize_virtual_chunk_access)
+    if not entries:
+        return None
+    return icechunk.containers_credentials(
+        {prefix: options.to_credential() for prefix, options in entries.items()}
+    )
